tests/wpt/tests/css/filter-effects/backdrop-filter-important.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS Filters: !important flag parsing</title>
    <link rel="help" href="https://crbug.com/340100643">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="target" style="backdrop-filter: blur(24px) !important">Some text.</div>
  </body>
  <script>
test(() => {
  assert_equals(target.style.backdropFilter, 'blur(24px)');
});
  </script>
</html>
